Moar Mac sanity checks; rename archive

+++ build-dr.py
@@ -1,6 +1,7 @@
 import subprocess
 import os
 import shutil
+import sys
 
 DEST_DIRECTORY = '.'
 
@@ -9,7 +10,7 @@ if os.path.isdir("upx"):
 else:
     upx_string = ""
 
-if os.path.isdir("build"):
+if os.path.isdir("build") and not sys.platform.find("mac") and not sys.platform.find("osx"):
     shutil.rmtree("build")
 
 subprocess.run(" ".join(["pyinstaller DungeonRandomizer.spec ",
